pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iamsentry"
version = "0.4.0"  # Keep in sync with IAMSentry/constants.py
description = "GCP IAM Security Auditor and Remediation Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Documo DevOps", email = "devops@documo.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
keywords = ["gcp", "iam", "security", "audit", "remediation", "google-cloud"]

dependencies = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "google-cloud-secret-manager>=2.0.0",
    "PyYAML>=6.0",
    "pydantic>=2.0.0",
    "schedule>=1.0.0",
    "httplib2>=0.20.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.0.0",
]
dashboard = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "aiofiles>=23.0.0",
    "jinja2>=3.0.0",
    "python-multipart>=0.0.6",
]

[project.scripts]
iamsentry = "IAMSentry.cli:cli_main"
iamsentry-legacy = "IAMSentry.manager:main"
iamsentry-dashboard = "IAMSentry.dashboard.server:main"

[project.urls]
Homepage = "https://github.com/documo/iamsentry"
Documentation = "https://github.com/documo/iamsentry#readme"
Repository = "https://github.com/documo/iamsentry.git"
Issues = "https://github.com/documo/iamsentry/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["IAMSentry*"]

[tool.setuptools.package-data]
IAMSentry = ["*.yaml", "*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["IAMSentry"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
